// Boardlight/src/api/generated/models/shopIndexProductsResponse.ts
/**
 * Generated by orval v6.16.0 🍺
 * Do not edit manually.
 * Blueboard
 * OpenAPI spec version: v4.0.0
 */

export interface ShopIndexProductsResponse {
  id?: number;
  name?: string | null;
  description?: string | null;
  visible?: boolean;
  qrCodeActivated?: boolean;
  price?: number;
  quantity?: number;
  thumbnailUrl?: string | null;
  createdAt?: string;
  updatedAt?: string;
}